Refactored the users component in admin to be manage-user and likewise added manage-library component.
