+++ src/app/admin/admin-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { AdminGuard } from '../_guards/admin.guard';
 import { DashboardComponent } from './dashboard/dashboard.component';
-import { UsersComponent } from './users/users.component';
+import { ManageUsersComponent } from './manage-users/manage-users.component';
 
 const routes: Routes = [
   {path: '**', component: DashboardComponent, pathMatch: 'full'},
@@ -11,7 +11,7 @@ const routes: Routes = [
     canActivate: [AdminGuard],
     children: [
       {path: '/dashboard', component: DashboardComponent},
-      {path: '/users', component: UsersComponent}
+      {path: '/users', component: ManageUsersComponent}
     ]
   }
 ];
