+++ src/app/admin/admin.module.ts
@@ -1,15 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AdminRoutingModule } from './admin-routing.module';
-import { UsersComponent } from './users/users.component';
-import { ToastrModule } from 'ngx-toastr';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { NgbNavModule } from '@ng-bootstrap/ng-bootstrap';
+import { ManageLibraryComponent } from './manage-library/manage-library.component';
+import { ManageUsersComponent } from './manage-users/manage-users.component';
 
 
 
 @NgModule({
-  declarations: [UsersComponent, DashboardComponent],
+  declarations: [ManageUsersComponent, DashboardComponent, ManageLibraryComponent],
   imports: [
     CommonModule,
     AdminRoutingModule,
