+++ src/app/admin/manage-library/manage-library.component.html
@@ -1,0 +1,3 @@
+<h2>Libraries</h2>
+
+<button class="btn btn-primary" (click)="addFolder('')">Add Folder</button>
