+++ src/app/admin/manage-users/manage-users.component.html
@@ -1,10 +1,4 @@
 
-<button class="btn btn-primary" (click)="addFolder('')">Add Folder</button>
-
-
-<div class="container">
-    <h2></h2>
-</div>
 
 <h2>Members:</h2>
 <div class="container">
